+++ pandapower/toolbox.py
@@ -436,9 +436,9 @@ def convert_format(net):
         net._ppc2pd_lookups = {"bus": None,
                             "ext_grid": None,
                             "gen": None}
-    if not "_is_elements" in net and "_is_elems" in net:
-        net["_is_elements"] = copy.deepcopy(net["_is_elems"])
-        net.pop("_is_elems", None)
+    if not "_is_elements" in net and "__is_elements" in net:
+        net["_is_elements"] = copy.deepcopy(net["__is_elements"])
+        net.pop("__is_elements", None)
 
     return net