+++ report/latex_postprocessing.py
@@ -6,7 +6,7 @@
 
 
 latex_file_path = "..\\doc\\_build\\latex\\pandapower.tex"
-version = "1.2.0"
+version = "1.2.2"
 
 class AlreadyProcessedException(Exception):
     pass